refactor: rename auth_impl.dart to auth_repository_impl.dart for clarity
